Refactor tests
Add common base class for chiasma tests, with a setup method. Move
chiasma tests to separate folder.
