+++ tests/unit/chiasma_tests/chiasma_validation_deploy_tests.py
@@ -1,39 +1,12 @@
-import unittest
 import os
 from unittest import skip
-from pyfakefs import fake_filesystem
-from release_ccharp.snpseq_workflow import SnpseqWorkflow
-from release_ccharp.snpseq_paths import SnpseqPathActions
-from release_ccharp.apps.chiasma import Application
-from tests.unit.utility.fake_os_service import FakeOsService
-from tests.unit.utility.fake_windows_commands import FakeWindowsCommands
+from tests.unit.chiasma_tests.base import ChiasmaBaseTests
 
 
-class ChiasmaValidationDeployTests(unittest.TestCase):
+class ChiasmaValidationDeployTests(ChiasmaBaseTests):
     def setUp(self):
-        config = {
-            "root_path": r'c:\xxx',
-            "git_repo_name": "chiasma",
-            "exe_file_name_base": "Chiasma",
-            "owner": "GitEdvard"
-        }
-        branch_provider = FakeBranchProvider()
-        wf = SnpseqWorkflow(whatif=False, repo="chiasma")
-        wf.config = config
-        wf.paths.config = config
-        wf.paths.branch_provider = branch_provider
-        self.filesystem = fake_filesystem.FakeFilesystem()
-        os_service = FakeOsService(self.filesystem)
-        self.os_module = os_service.os_module
-        self.chiasma = Application(wf, branch_provider, os_service,
-                                   FakeWindowsCommands(self.filesystem), whatif=False)
+        self.base_setup()
         self.file_builder = FileSystemBuilder(self.chiasma, self.filesystem)
-
-        path_actions = SnpseqPathActions(
-            whatif=False, snpseq_path_properties=self.chiasma.path_properties,
-            os_service=os_service
-        )
-        path_actions.generate_folder_tree()
 
     def test_create_shortcut__with_latest_empty__something_is_copied_to_latest(self):
         self.chiasma.validation_deployer.create_shortcut()
@@ -231,13 +204,6 @@ class ChiasmaValidationDeployTests(unittest.TestCase):
         latest = r'c:\xxx\chiasma\uservalidations\latest'
         dir_objects = [o for o in self.chiasma.os_service.listdir(latest)]
         self.assertEqual(2, len(dir_objects))
-
-
-class FakeBranchProvider:
-    def __init__(self):
-        self.candidate_version = "1.0.0"
-        self.latest_version = "latest-version"
-        self.candidate_branch = "release-1.0.0"
 
 
 class FileSystemBuilder:
